move prometheus to its own file
