+++ pkg/api/router.go
@@ -3,10 +3,7 @@ package api
 import (
 	"context"
 	"net/http"
-	"strconv"
-	"time"
 
-	"github.com/go-chi/chi/middleware"
 	"github.com/go-chi/chi/v5"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
@@ -23,21 +20,6 @@ var responseSize = promauto.NewHistogramVec(prometheus.HistogramOpts{
 	Help:    "Bytes returned",
 	Buckets: prometheus.ExponentialBucketsRange(1000, 100_000_000, 20),
 }, []string{"route"})
-
-func PrometheusMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		start := time.Now()
-
-		ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
-		next.ServeHTTP(ww, r)
-
-		duration := time.Since(start)
-		routeName := chi.RouteContext(r.Context()).RoutePattern()
-
-		latency.WithLabelValues(routeName, strconv.Itoa(ww.Status())).Observe(duration.Seconds())
-		responseSize.WithLabelValues(routeName).Observe(float64(ww.BytesWritten()))
-	})
-}
 
 type ScratchDataAPI interface {
 	Select(w http.ResponseWriter, r *http.Request)
